--- src/components/visualizer-component/insertionSort.js
@@ -1,5 +1,0 @@
-const insertionSort = (array) => {
-  return;
-}
-
-export default insertionSort;
