+++ src/components/visualizer-component/sorting-functions/selectionSort.js
@@ -1,0 +1,9 @@
+// Importing node modules
+import {v4 as uuid} from 'uuid';
+import Element from '../../element-component/Element';
+
+const selectionSort = (array, stopSort, setElementsArray, speed) => {
+  return;
+}
+
+export default selectionSort;
